feat: WIP first pass internal call fuzzer

+++ barretenberg/cpp/src/barretenberg/vm2/tracegen/execution_trace.cpp
@@ -456,7 +456,11 @@ void ExecutionTraceBuilder::process(
             trace.set(row,
                       { {
                           { C::execution_next_pc,
-                            ex_event.before_context_event.pc + ex_event.wire_instruction.size_in_bytes() },
+                            // TODO(MW): The fuzzer discovered that the below can overflow: context.get_pc() is set as a
+                            // uint32 and the internal call stack gathers that value, whereas below added in the field.
+                            // Temp soln:
+                            static_cast<uint32_t>(ex_event.before_context_event.pc +
+                                                  ex_event.wire_instruction.size_in_bytes()) },
                       } });
 
             // Along this function we need to set the info we get from the #[EXEC_SPEC_READ] lookup.
